Version 0.7.4: Fixed INCLUDE bug
When INCLUDEing a transcript file, the title/author was added to
the main title/author, creating the wrong title/author.  The
solution was to added a heading to the transcript, then INCLUDE
only the contents, using :contents-only t.

Also placed export settings in the transcript into a Export
heading that is not exported.